requirements.txt
requests
beautifulsoup4
lxml
python-dateutil
selenium
instructor[google-genai]
pydantic
python-dotenv
google-genai
